requirements.txt
tkcalendar==1.6.1
mysql-connector-python==8.0.30
Pillow==9.2.0